xpdf 4.00
git-svn-id: svn://tug.org/texlive/trunk/Build/source@45031 c570f23f-e606-0410-a88d-b1316a301751

+++ libs/xpdf/xpdf-src/goo/FixedPoint.cc
@@ -16,6 +16,7 @@
 #pragma implementation
 #endif
 
+#include "gmempp.h"
 #include "FixedPoint.h"
 
 #define ln2 ((FixedPoint)0.69314718)
@@ -113,7 +114,7 @@ GBool FixedPoint::divCheck(FixedPoint x, FixedPoint y, FixedPoint *result) {
       z >= ((FixPtInt64)1 << 31) || z < -((FixPtInt64)1 << 31)) {
     return gFalse;
   }
-  result->val = z;
+  result->val = (int)z;
   return gTrue;
 }
